+++ newsletter-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao newsletter bundle change log
 
+### DEV
+
+ * Trigger all the callbacks in the toggleVisibility() methods (see contao/core-bundle#756).
+
 ### 4.4.0-beta1 (2017-05-05)
 
  * Optimize the element preview height (see contao/core-bundle#678).